bug 2496: add 'edit' link to authority search results
When search for an authority record from the bib
record editor, a new link is added to each
authority result to edit the authority record.  This
allows a cataloger to do the following:

1. Start editing a bib record
2. Click on the '...' to select an authorized heading
   for an authority-controlled field.
3. Look at the results in the pop-up window and
   see that an authority record needs correction.
4. Click on the 'edit authority' link.  This allows
   the authority to be revised.  Note that this
   takes place in the popup window, not the originally
   bib editing window.
5. Save the authority record, then close the popup
   window and redo the heading search.

For various reasons, this not a perfect workflow; this
patch is designed to satisfy an immediate requirement
but needs to be reworked in 3.2.

Signed-off-by: Joshua Ferraro <jmf@liblime.com>