bug 2520: don't change item-level requests to bib-level
Prior to this fix, changing the priority of any hold requests
in the queue of requests for a bib would convert
all item-level requests to title-level requests.

As part of this fix, improved display of the holds queue
so that item-level and bib-level requests can be more
clearly distinguished.

Documentation note: new screenshot of list of requests for
a bib in the staff interface.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>