bug 2523: add columns to holds queue report
* added author, ccode, location, and item enumchron
* sorted by collection, locatoin, call number, author, title

As part of this patch, commented out the JavaScript tablesorter,
which could cause the browerer to lock up if the picklist
contains more than a couple hundred items to pull.
Some sort of pagination is clearly required, but since
it is necessary for customers to be able to print out
the picklist, there still needs to be an option to
see the complete list for a branch.

This patch includes the final component of the fix
for bug 2331.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>