bug 2568: set default value of intranetuserjs to blank
The intranetuserjs syspref allows a Koha administrator
to insert arbitrary JavaScript into the staff page header.
Now that this syspref is functional again, this patch
changes the installation scripts so that the default
value is blank - any JavaScript we ship with Koha should
be in the templates, not in a syspref.

Documentation changes: none per se.  However, I do want to
point out that the intranetuserjs syspref is a bit of a
dangerous one - enter a value with incorrect syntax, and
you could make it difficult to access the staff interface -
and should probably be marked as an advanced option.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>