bug 2585: fix search links for names (XSLT OPAC display)
Fix malformed OPAC search syntax that caused this bug:

When using the default stylesheet for MARC21 XSLT OPAC details display, the
search links for name headings (from 100, 700, 710, 710, 111, and 711) that are
linked to authority records don't work.  When the user clicks on such a link,
they are directed to the advanced search form instead of the search results.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>