bug 2608: let offline circ processing work in background
If an offline circulation file is large, process_koc.pl
could time out before processing all of the transactions.

To prevent this, the processing can now be run in the
background, with an AJAX loop to check the job's
status.

As a consequence of the patch, the UI for uploading offline circ files has
changed slightly.  One must select a file, click the "upload file" button, then
click the "process offline circulation file" button.

Also removed the now-superfluous upload_koc.pl, added the warnings pragma,
and fixed a typo in the template.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>