bug 2625
Using the barcode as an index for assigning values to an array creates problems if
the numeric value of the barcode is sufficiently large to consume enough memory to
starve other processes. This patch changes return_failed from an indexed array into
an associative one, using the barcode as the key.

Signed-off-by: Ryan Higgins <rch@liblime.com>
Signed-off-by: Galen Charlton <galen.charlton@liblime.com>