bug 2825: improving readability of output messages in updatedatabase.pl
Two recent changes to updatedatabase.pl don't include newlines at the
end of their status messages. This patch adds them. This is only useful
if you're running this from the command line.

Note: there is an opportunity here to extract some of the simlar logic
from each of these blocks to a function.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>