Modify GoogleJacket display
On detail page, when GoogleJacket is enabled,
book cover is retrieved from Google. This image
is a link to GoogleBook services: full text,
book localisation in a bookstore, etc.

This patch removes the link on book cover and
add underneath a text 'Source Google' with a
link to GoogleBook services.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>