Script to repair MARC21 leader/09.
Acquisitions process seems to be adding records with incorrect
representation of the MARC encoding in leader/09.  It should be
'a' meaning UTF-8, for all Koha's internalized records, but in
many cases it appears blank (for MARC-8).  This script diagnoses
and repairs the value in the leader, depending on runtime options.

The symptom of this problem is that high-value UNICODE characters
in the record will cause Koha to crash whenever it tries to parse
the MARCXML, giving a "Wide character" fatal.  While we work on
fixing the input, this script will fix the existing data.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>