new author test for syntax errors in HTML templates
Added a new author test to look for syntax errors
in the HTML templates.  Run by doing

prove xt/author/valid-templates.t

from the root of a Koha source tree.

This test will help catch the following errors:

* unbalanced TMPL_IF, TMPL_UNLESS, and TMPL_LOOP contructs, e.g.,
  cases where a TMPL_IF is not closed by a /TMPL_IF.
* references to nonexistant include files
* syntax errors within a HTML::Template::Pro tag

This test currently ignores errors related to TMPL_IF EXPR, which
is currently deprecated for use in Koha -- this may be made
stricter in the future.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>