Standalone branch selection
selectbranchprinter now posts to itself to set the branch.  That also means that
other pages can set the branch by linking to selectbranchprinter like:
    /cgi-bin/koha/circ/selectbranchprinter.pl?branch=MAIN

We use HTTP_REFERER to allow the user to continue back to wherever they linked from,
defaulting to circulation.pl (the former POST target).

Note that all "branchprinter" functions are basically useless as the data is not
properly handled in the session storage.  Chris C. expects this to be addressed
with later patches for HLT.  There appear to be no 3.0 users of this intended
functionality yet.

Includes squashed patch to fix tmpl and to automatically redirect if there are no
POST params to be recycled.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>