Modified fines.pl to not throw warning if --out not specified
It appears, in Ryan's patch, that he wants to throw a warning to the log if
the directory specified in --out is not present.  (Further messages will
be given when the open-or-die occurs a few lines later.)  However, it was
throwing the warning if --out was not specified at all, which is
undesirable.  This patch modifies that bit to check for the presence of
whatever directory is going to be used, either --out, ENV{TMPDIR}, or /tmp.
As before, if the write to the directory fails for any reason--including
its' non-existence--that is handled later, but this message will help
inform the troubleshooter.

Signed-off-by: Galen Charlton <galen.charlton@liblime.com>