comment out styles in gb_styles.css
These were added in the labels rework, but do not belong here, since
gb_styles.css should references only CSS classes responsible for
displaying the greybox itself, not the content inside it.  In particular,
the generic table styles were causing the icons added by tablesorter
to be tiled in the opac-detail template.

To finish this work, the styles should be taken out entirely and
moved to whichever lables template they were supposed to affect

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>