Handle null-or-empty to Charset::StripNonXmlChars
When rebuild_zebra.pl is run from cron, there is an occasional error
of the form:

Use of uninitialized value $str in substitution (s///) at /home/ebpl/kohaclone/C4/Charset.pm line 304.

This error is occuring when the string that is fed to Charset::StripNonXmlChars
is null or undefined, for some reason.

This fix will handle the null-or-empty condition, and thus suppress the error.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>