Fix for Bug 3730 - Wrong ISO 639-2 Language Code for Portuguese in database
[Patch originally by Ricardo Dias Marques]

Bug 3730 - Wrong ISO 639-2 Language Code for Portuguese in database
http://bugs.koha.org/cgi-bin/bugzilla3/show_bug.cgi?id=3730

For 3.0.x head of the Git branch, the ISO 639-2 Language Code for
Portuguese in Koha database is wrong.

Specifically, in the table "language_rfc4646_to_iso639", when the
field "rfc4646_subtag" has the value "pt", the field "iso639_2_code"
should have the value of "por" and NOT the current one of "pol"
("pol" refers to "Polish" as correctly also appears for the code "pl",
in the row above the "pt" one).

In my test setups, this showed some serious sequences for people
that choose the Portuguese language in the Web Installer, namely:
- very weird look of Staff client with missing Header and Footer,
that renders the Staff client unusable
- Portuguese translation apparently not getting updated for the OPAC
(for people that do an Upgrade from a previous version)

This patch is for the 3.0.x branch (I tested it with that branch).

For the master branch, I don't have a database setup to create and
test a similar patch but Chris Cormack nicely volunteered to do that.
Thanks Chris!

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>