Bug 3750 Modify XSLT to display subfield n of 111/711 and remove period after 110/710
If a 111/711 uses a subfield n (Number of part/section/meeting), XSLT wasn't displaying it which made for an incomplete display.

	e.g., International Conference on Concurrency Theory 1999 : Eindhoven, Netherlands)
	instead of 10th International Conference on Concurrency Theory 1999 : Eindhoven, Netherlands)

I've added it to show before subfield a if present.  I've also removed the period that XSLT puts in the results and display after the last occurrence of a 110/710.  It's fairly standard practice for these entries to have a period in the MARC record, so the XSLT display was showing a double period.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>