Fix for MSWin32
A tweak to restore MSWin32 installation capability.

NOTE: Native MSWin32 support in Koha is for Strawberry Perl 5.10.1.0 or later.
There are some issues with GDBM_File.pm which prevent Koha from running without a bit of
a workaround. This will be resolved in the January 2010 release of Strawberry Perl. More
can be read concerning this in the INSTALL.MSWin32 file which will be forth-coming.

NOTE: There will be several other minor changes to other portions of the Koha code base
in order to fix misc issues when Koha is installed on MSWin32.