Bug 3890: Nothing in Type Pull Down on Purchase Suggestions.
This patch adds a description of the itype next to the image that i suppose to appear in the selection list.

Fixed the path of the image in opac-suggestions.pl, but I'm not sure if you can actually display an image in a <SELECT> list option.

Also removed an extra </label> closing tag.