(bug #3584) detect ccl queries
This patch C4::Search::buildQuery to detect ccl queries, and let zebra to parse them.
And set a default index "kw" if not specified.
This improve the detection of ccl queries, and do not duplicate the "ccl=" value.
Adding = as an index sign
Followup to 3037ff9e81149615f94d3c4ff0e5b4c5240f3ce9

Signed-off-by: Henri-Damien LAURENT <henridamien.laurent@biblibre.com>