Fix for Bug 3991, Circulation search submit button wrapping in WebKit browsers
If CircAutocompl is ON, the circ search form has a hidden <div> for use in
displaying live search results. For some reason WebKit browsers (Safari and
Chrome) treat this hidden element as a something that should create a line
break like any other block-level element. Moving the submit button to
before that hidden <div> fixes the problem without disrupting the display
of the live search results.