Bug 4057 Fix spacing in XSLT OPAC display for 245 subfield f
If a 245 has a subfield f (or any of the displayed fgknps subfields after a, b,
or h) there is no spacing and the display runs together.  This happens in both
the XSLT results list and the XSLT individual title display.  I've added a
space between the a, b, h displays and the beginning of the fgknps set.