Bugfix: Make smart-rules.tmpl display zero values in fields.
Since TMPL_IFs evaluate false if the template variable is 0,
smart-rules.tmpl was not displaying zero values in fields where
the db field value was indeed zero which leads to some small confusion.

This patch removes the TMPL_IFs since those fields will always be
populated with some value even if it is zero.

Signed-off-by: Galen Charlton <gmcharlt@gmail.com>